+++ components/detail.jsx
@@ -3,8 +3,8 @@ var React = require('react');
 var Detail = React.createClass({
     render: function () {
         return (
-        <div class="detail">
-        <img src={this.post.url} />
+        <div className="detail">
+        <img src={this.props.url} />
         <ul>
         <li><strong>Title:</strong>{this.props.title}</li>
         <li><strong>Description:</strong>{this.props.description}</li>
@@ -14,6 +14,8 @@ var Detail = React.createClass({
      )
     }
 });
+
+module.exports = Detail;
