+++ components/edit.jsx
@@ -1,0 +1,28 @@
+/*var React = require('react');
+
+var EditAdd = React.createClass({
+render function() {
+    return (
+    <form id="detailForm">
+    <input id="objectId" type="hidden" value="{this.props.objectId}"/>
+    <div className="form-group">
+    <label for="title">Title</label>
+    <input type="text" className="form-control" id="title" placeholder="Title" value="{title}">
+    </div>
+    <div className="form-group">
+    <label for="url">Image URL</label>
+    <input type="text class="form-control" id="url" placeholder="Image URL" value="{this.props.url}">
+    </div>
+    <div className="form-group">
+    <label for="description">Description</label>
+    <textarea id="description" className="form-control" rows="3" placeholder="Description">{this.props.description}</textarea>
+    </div>
+    <button className="btn btn-default" type="submit">Submit</button>
+    </form>
+    )
+  }
+});
+
+module.exports = EditAdd;*/
+
+
