+++ js/router.jsx
@@ -2,10 +2,16 @@ var Backbone = require('./backbone-parse');
 var Post = require('./models/post');
 var Posts = require('./collections/posts');
 //var mainTemplate = require('./templates/main.html');
-var detailTemplate = require('./templates/detail.html');
-var addEditTemplate = require('./templates/addEdit.html');
-//var React = require('./react');
-//var Main = require('./')
+//var detailTemplate = require('./templates/detail.html');
+//var addEditTemplate = require('./templates/addEdit.html');
+var React = require('react');
+var ReactDOM = require('react-dom');
+var Main = require('./components/main.jsx');
+var ReactDOM = require('react-dom');
+var Detail = require('./components/detail.jsx');
+var ReactDOM = require('react-dom');
+var Edit = require('./components/edit.jsx');
+
 
 var Router = Backbone.Router.extend({
   initialize: function () {
@@ -20,12 +26,12 @@ var Router = Backbone.Router.extend({
   index: function () {
     Posts.fetch({
       success: function (posts) {
-       var data = posts.toJSON();
-          
+      //var html = mainTemplate({'data': posts.toJSON()});
+      //$("#container").html(html);
           var data = posts.toJSON();
           ReactDOM.render(<Main />, document.getElementById('container'));
       }
-        Main=(data);
+      Main=(data);
     });
   }
 });
@@ -35,18 +41,20 @@ var router = new Router();
 router.on('route:post', function (objectId) {
   var post = Posts.get(objectId);
   var html = detailTemplate(post.toJSON());
-  $("#container").html(html);
+  //$("#container").html(html);
+  ReactDOM.render(<Detail />, document.getElementById('container'));
 });
 
 router.on('route:add', function () {
-  var html = addEditTemplate({});
-  $("#container").html(html);
+ReactDOM.render(<addEditForm />, document.getElementById('container'));
+//var html = addEditTemplate({});
+  //$("#container").html(html);
 });
 
 router.on('route:edit', function (objectId) {
   var post = Posts.get(objectId);
-  var html = addEditTemplate(post.toJSON());
-  $("#container").html(html);
+  //var html = addEditTemplate(post.toJSON());
+  //$("#container").html(html);
 });
 
 $('body').on('click', 'a', function (e){
